Reemplazar algoMon muerto
Ya se puede reemplazar un algoMon cuando muere.

+++ src/fiuba/algo3/modelo/estados/Quemado.java
@@ -30,7 +30,7 @@ public final class Quemado extends EstadoAlterado {
 		this.getEstadoAnterior().accionRealizada();
 
 		try {
-			this.quitarVida(this.vidaMaxima * Quemado.getPorcentajeVidaQuitada());
+			this.quitarVida((int) (this.vidaMaxima * Quemado.getPorcentajeVidaQuitada()));
 			throw new AlgoMonRecibeDañoQuemadura(/*"¡El algoMon recibe daño de la quemadura!"*/);
 		} catch(AlgoMonMurio e) {
 			throw new AlgoMonMurioPorQuemadura(/*"¡El algoMon murió a causa de la quemadura!"*/);
